fixed dropdown props; updated venue spaces list and icons usage
