+++ src/utils/room.ts
@@ -2,15 +2,6 @@ import Bugsnag from "@bugsnag/js";
 import Video from "twilio-video";
 
 import { ParticipantWithUser } from "types/rooms";
-import { VenueTemplate } from "types/venues";
-
-import RoomIconArtPiece from "assets/icons/icon-room-artpiece.svg";
-import RoomIconAuditorium from "assets/icons/icon-room-auditorium.svg";
-import RoomIconBurnBarrel from "assets/icons/icon-room-burnbarrel.svg";
-import RoomIconConversation from "assets/icons/icon-room-conversation.svg";
-import RoomIconExperience from "assets/icons/icon-room-experience.svg";
-import RoomIconMap from "assets/icons/icon-room-map.svg";
-import RoomIconMusicBar from "assets/icons/icon-room-musicbar.svg";
 
 export interface GetExternalRoomSlugProps {
   venueName: string;
@@ -45,47 +36,3 @@ export const logIfCannotFindExistingParticipant = (
     );
   }
 };
-
-export interface VenueSpace {
-  text: string;
-  template?: VenueTemplate;
-  icon: string;
-}
-
-export const venueSpacesList: VenueSpace[] = [
-  {
-    text: "Conversation Space",
-    icon: RoomIconConversation,
-    template: VenueTemplate.conversationspace,
-  },
-  {
-    text: "Auditorium",
-    icon: RoomIconAuditorium,
-    template: VenueTemplate.audience,
-  },
-  {
-    text: "Music Bar",
-    icon: RoomIconMusicBar,
-    template: VenueTemplate.jazzbar,
-  },
-  {
-    text: "Burn Firebarrel",
-    icon: RoomIconBurnBarrel,
-    template: VenueTemplate.firebarrel,
-  },
-  {
-    text: "Art Piece",
-    icon: RoomIconArtPiece,
-    template: VenueTemplate.artpiece,
-  },
-  {
-    text: "Experience",
-    icon: RoomIconExperience,
-    template: VenueTemplate.zoomroom,
-  },
-  {
-    text: "Map",
-    icon: RoomIconMap,
-    template: VenueTemplate.partymap,
-  },
-];
